src/main/java/com/marcelopazzo/fragmentapplication/MainActivity.java
package com.marcelopazzo.fragmentapplication;

import android.os.Bundle;

import com.actionbarsherlock.app.SherlockFragmentActivity;
import com.actionbarsherlock.view.MenuItem;

public class MainActivity extends SherlockFragmentActivity {

	@Override
	protected void onCreate(Bundle savedInstanceState) {
		super.onCreate(savedInstanceState);
		setContentView(R.layout.activity_main);
	}

	@Override
	public boolean onOptionsItemSelected(MenuItem item) {
		if (item.getItemId() == android.R.id.home)
		{
			finish();
			return false;
		} else {
			return super.onOptionsItemSelected(item);
		}
	}

}
